+++ packages/user/src/index.ts
@@ -45,6 +45,8 @@ declare module "@dzangolab/fastify-config" {
           adminSignUp?: typeof userHandlers.adminSignUp;
           canAdminSignUp?: typeof userHandlers.canAdminSignUp;
           changePassword?: typeof userHandlers.changePassword;
+          disable?: typeof userHandlers.disable;
+          enable?: typeof userHandlers.enable;
           me?: typeof userHandlers.me;
           updateMe?: typeof userHandlers.updateMe;
           users?: typeof userHandlers.users;
@@ -101,10 +103,8 @@ export { default as validatePassword } from "./validator/password";
 
 export * from "./constants";
 
-export type {
-  EmailVerificationRecipe,
-  ThirdPartyEmailPasswordRecipe,
-} from "./supertokens/types";
+export type { EmailVerificationRecipe } from "./supertokens/types/emailVerificationRecipe";
+export type { ThirdPartyEmailPasswordRecipe } from "./supertokens/types/thirdPartyEmailPasswordRecipe";
 export type {
   AuthUser,
   ChangePasswordInput,